fixes bug where errors weren't showing properly - removes navigating responsibilty out of useSearchDao hook - updates useAddress hook to support search hook
